% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fit.R
\name{read.fit}
\alias{read.fit}
\title{Function to read {length, weight}.fit files}
\usage{
read.fit(fit.file, verbose = FALSE, rep = NULL,
  overall.composition.plot = FALSE, fit = FALSE, plot = TRUE,
  plot.ctl = list(xlabel = "Length(cm)", dir = "h", line.wdth = 0.5, Ncols =
  2, lincol = "#FF3333", fillcol = "#6699CC", nbrks = 3))
}
\description{
Function to read {length, weight}.fit files
}
