% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.par.R
\name{read.par}
\alias{read.par}
\title{by Simon Hoyle June 2008
This is only partly built. Needs to be extended so it gets the whole par file into an object. Then do the same for write.par...
Nick : added sections explicitly for # tag flags; # tag fish rep; # tag fish rep group flags; # tag_fish_rep active flags; # tag_fish_rep target; # tag_fish_rep penalty
NMD 22/06/12 - allow instance of no tag flags or tag pars at all (e.g. striped marlin)
YT : 27/02/17 : Upgraded for multi species/sex model
YT : 10/06/2017 : made it compatible to para file version 1042}
\usage{
read.par(par.file, verbose = TRUE)
}
\arguments{
\item{par.file}{charactor file name of par file}

\item{verbose}{if TRUE verbose outputs will be made}
}
\description{
by Simon Hoyle June 2008
This is only partly built. Needs to be extended so it gets the whole par file into an object. Then do the same for write.par...
Nick : added sections explicitly for # tag flags; # tag fish rep; # tag fish rep group flags; # tag_fish_rep active flags; # tag_fish_rep target; # tag_fish_rep penalty
NMD 22/06/12 - allow instance of no tag flags or tag pars at all (e.g. striped marlin)
YT : 27/02/17 : Upgraded for multi species/sex model
YT : 10/06/2017 : made it compatible to para file version 1042
}
\author{
Simon Hoyle
}
