% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.rep.R
\name{read.rep}
\alias{read.rep}
\title{Function to read plot.rep file}
\usage{
read.rep(rep.file, verbose = FALSE, DEBUG = FALSE)
}
\arguments{
\item{rep.file}{CHARACTER file name of plot.rep file}

\item{verbose}{LOGICAL if TRUE produce verbose outputs on screen}
}
\description{
Function to read plot.rep file
}
